+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/unnecessaryparentheses/InputUnnecessaryParentheses15Extensions.java
@@ -1,0 +1,36 @@
+// someexamples of 1.5 extensions
+package com.puppycrawl.tools.checkstyle.checks.coding.unnecessaryparentheses;
+
+@interface MyAnnotation1 {
+    String name();
+    int version();
+}
+
+@MyAnnotation1(name = "ABC", version = 1)
+public class InputUnnecessaryParentheses15Extensions
+{
+
+}
+
+enum Enum2
+{
+    A, B, C;
+    Enum2() {}
+    public String toString() {
+        return ""; //some custom implementation
+    }
+}
+
+interface TestRequireThisEnum
+{
+    enum DAY_OF_WEEK
+    {
+        SUNDAY,
+        MONDAY,
+        TUESDAY,
+        WEDNESDAY,
+        THURSDAY,
+        FRIDAY,
+        SATURDAY
+    }
+}
